+++ CHANGES.md
@@ -9,6 +9,7 @@
 * add REST API endpoints courtesy of @limikael
 * simplified and removed lots of code for viewing theme details, use core when available. ;-)
 * added CSS file to help display theme view details, waiting for [#36923](https://core.trac.wordpress.org/ticket/36923) for it to work
+* refactored `get_remote_{plugin|theme}_meta()` to `get_remote_repo_meta()` as it was in 4 different places :P
 
 #### 5.4.1
 * get tags for themes to rollback even if no updates are available. I was overzealous in cutting remote API calls.